requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
